requirements.txt
requests
requests[security]
beautifulsoup4
python-telegram-bot
# For summarization, this bot uses the Hugging Face Inference API (facebook/bart-large-cnn) 